requirements.txt
Django==1.10.4
dj-database-url==0.4.1
dj-static==0.0.6
django-compressor==2.1
django-toolbelt==0.0.1

facebook-sdk==2.0.0
gunicorn==19.6.0
psycopg2==2.6.2
static3==0.7.0
requests==2.12.5
requests-oauthlib==0.7.0
whitenoise==3.2.3
